+++ test/stats_test.dart
@@ -47,8 +47,10 @@ void main() {
       final numberOfIntervals = 10;
       final hist = stats.histogram(intervals: numberOfIntervals);
       var sum = hist[1].fold<num>(0.0, (sum, current) => sum + current);
-      expect(sum * (stats.max - stats.min) / numberOfIntervals,
-          closeTo(1.0, 1e-12));
+      expect(
+        sum * (stats.max - stats.min) / numberOfIntervals,
+        closeTo(1.0, 1e-12),
+      );
     });
     test('Total count (non-normalized histograms)', () {
       final hist = stats.histogram(normalize: false);
@@ -64,30 +66,31 @@ void main() {
     );
     test('Data', () {
       expect(
-          hist,
-          '# Intervals: 8\n'
-          '# Min: -1.9490799\n'
-          '# Max: 26.551828\n'
-          '# Mean: 10.168769\n'
-          '# StdDev: 5.3700258\n'
-          '# Median: 10.232570\n'
-          '# First Quartile: 6.1556008\n'
-          '# Third Quartile: 14.234971\n'
-          '# Interval size: 3.5626135\n'
-          '# Integrated histogram: 1.0000000000000002\n'
-          '#\n'
-          '# -------------------------------------------------------------\n'
-          '#     Range     Count    Prob. Density Func. \n'
-          '-1.9490799     0.0056138562     0.0058236414\n'
-          '1.6135336     0.022455425     0.020882819\n'
-          '5.1761471     0.056138562     0.048220974\n'
-          '8.7387606     0.058945490     0.071702644\n'
-          '12.301374     0.064559346     0.068657297\n'
-          '15.863988     0.056138562     0.042334124\n'
-          '19.426601     0.011227712     0.016809190\n'
-          '22.989215     0.0028069281     0.0042978905\n'
-          '26.551828     0.0028069281     0.00070764630\n'
-          '');
+        hist,
+        '# Intervals: 8\n'
+        '# Min: -1.9490799\n'
+        '# Max: 26.551828\n'
+        '# Mean: 10.168769\n'
+        '# StdDev: 5.3700258\n'
+        '# Median: 10.232570\n'
+        '# First Quartile: 6.1556008\n'
+        '# Third Quartile: 14.234971\n'
+        '# Interval size: 3.5626135\n'
+        '# Integrated histogram: 1.0000000000000002\n'
+        '#\n'
+        '# -------------------------------------------------------------\n'
+        '#     Range     Count    Prob. Density Func. \n'
+        '-1.9490799     0.0056138562     0.0058236414\n'
+        '1.6135336     0.022455425     0.020882819\n'
+        '5.1761471     0.056138562     0.048220974\n'
+        '8.7387606     0.058945490     0.071702644\n'
+        '12.301374     0.064559346     0.068657297\n'
+        '15.863988     0.056138562     0.042334124\n'
+        '19.426601     0.011227712     0.016809190\n'
+        '22.989215     0.0028069281     0.0042978905\n'
+        '26.551828     0.0028069281     0.00070764630\n'
+        '',
+      );
     });
   });
 }